requirements.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
albumentations==0.4.3
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.5.2
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.4.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
backcall==0.2.0
blinker==1.8.2
braceexpand==0.1.7
cachetools==5.5.0
carvekit-colab==4.1.0
causal-conv1d==1.2.0.post2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
clip-by-openai==1.1
cmake==3.31.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.9
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.4.0
debugpy==1.8.1
decorator==5.1.1
diffusers==0.12.1
dill==0.3.5.1
docker-pycreds==0.4.0
easydict==1.13
einops==0.3.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.115.6
fastcore==1.7.20
fasteners==0.19
fastjsonschema==2.21.1
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.4.0
filelock==3.16.1
fire==0.4.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.55.1
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.10.0
ftfy==6.2.3
future==1.0.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==3.21.0
grpcio==1.68.1
h11==0.14.0
hawkes==1.0.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
icecream==2.1.3
idna==3.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib_metadata==7.1.0
importlib_resources==6.3.1
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.5
itsdangerous==2.2.0
jax==0.4.13
jaxlib==0.4.13
jaxtyping==0.2.19
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
kornia==0.6.0
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.11.9
linkify-it-py==2.0.3
lit==18.1.8
loguru==0.7.2
lovely-numpy==0.2.13
lovely-tensors==0.1.18
lpips==0.1.4
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
ml-dtypes==0.2.0
modular-diffusion==0.0.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.13
narwhals==1.15.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numcodecs==0.12.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1687808301083/work
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
nvitop @ file:///nfs/home/wldn1677/nvitop
oauthlib==3.2.2
omegaconf==2.1.1
open3d==0.18.0
open_clip_torch==2.29.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.12
packaging==24.1
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
plotly==5.13.1
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==4.25.5
psutil==5.9.8
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
pyglet==2.0.18
Pygments==2.17.2
PyOpenGL==3.1.0
pyparsing==3.1.4
pyquaternion==0.9.9
pyrender==0.1.45
pyssim==0.7
python-dateutil==2.9.0.post0
python-multipart==0.0.19
pytorch-lightning==1.5.3
pytubefix==8.8.5
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
responses==0.18.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
safetensors==0.4.5
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.1
sentry-sdk==1.43.0
setproctitle==1.3.3
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
streamlit==1.40.1
sympy==1.12
taming-transformers-rom1504==0.0.6
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
test_tube==0.7.5
threadpoolctl==3.4.0
tifffile==2023.7.10
timm==0.6.7
tokenizers==0.12.1
toml==0.10.2
tomli==2.2.1
torch==1.12.1+cu113
torch-fidelity==0.3.0
torchaudio==0.12.1+cu113
torchcde==0.2.5
torchdiffeq==0.2.3
torchmetrics==1.5.2
torchsde==0.2.6
torchvision==0.13.1+cu113
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.2
trampoline==0.1.2
transformers==4.22.2
trimesh==4.5.3
triton==2.0.0
typeguard==4.4.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.3
urwid==2.6.16
uvicorn==0.32.1
wandb==0.16.5
watchdog==4.0.2
wcwidth==0.2.13
webdataset==0.2.5
websockets==13.1
Werkzeug==3.0.2
widgetsnbextension==4.0.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.15.2
zarr==2.16.1
zipp==3.18.1
